requirements.txt
streamlit
openai
google-generativeai
PyMuPDF
Pillow
pytesseract
python-dotenv